votacionController-votacionhtml-listacandidato-listausuarios-header home

+++ TPEvaluativo_Grupo12/src/main/java/ar/edu/unju/fi/controller/VotacionController.java
@@ -19,13 +19,13 @@ public class VotacionController {
 	@GetMapping("/votacion")
 	public String getVotacionPage(Model model) {
 		int totalvotos=0;
-		int porcentaje =0;
+		double porcentaje=0;
 		for(Candidato candid : listacandidatos.getCandidatos()) {
 			totalvotos=totalvotos+candid.getCantVotos();
 		}
 		for(int i=0;i<listacandidatos.getCandidatos().size();i++) {
 			int votoindividual = listacandidatos.getCandidatos().get(i).getCantVotos();
-			porcentaje=((100*votoindividual)/totalvotos);
+			porcentaje=((100.0*votoindividual)/totalvotos);
 			listacandidatos.getCandidatos().get(i).setPorcentaje(porcentaje);
 			LOGGER.info("--"+listacandidatos.getCandidatos().get(i).getNombre()+listacandidatos.getCandidatos().get(i).getPorcentaje());
 		}
